a lot of fixes & added friend requests' endpoints

+++ src/schemas/gd/account.ts
@@ -18,11 +18,11 @@ export const loginGJAccountSchema = z.object({
 export type LoginGJAccountInput = z.infer<typeof loginGJAccountSchema>;
 
 export const updateGJAccSettingsSchema = z.object({
-    accountID: z.coerce.number(),
+    accountID: z.coerce.number().int(),
     gjp2: z.string(),
-    mS: z.coerce.number(),
-    frS: z.coerce.number(),
-    cS: z.coerce.number(),
+    mS: z.coerce.number().int(),
+    frS: z.coerce.number().int(),
+    cS: z.coerce.number().int(),
     yt: z.string(),
     twitter: z.string(),
     twitch: z.string(),
@@ -32,7 +32,7 @@ export const updateGJAccSettingsSchema = z.object({
 export type UpdateGJAccSettingsInput = z.infer<typeof updateGJAccSettingsSchema>;
 
 export const backupGJAccountNewSchema = z.object({
-    accountID: z.coerce.number(),
+    accountID: z.coerce.number().int(),
     gjp2: z.string(),
     saveData: z.string(),
     gameVersion: z.coerce.number().int(),
@@ -43,7 +43,7 @@ export const backupGJAccountNewSchema = z.object({
 export type BackupGJAccountNewInput = z.infer<typeof backupGJAccountNewSchema>;
 
 export const syncGJAccountNewSchema = z.object({
-    accountID: z.coerce.number(),
+    accountID: z.coerce.number().int(),
     gjp2: z.string(),
     secret: z.string()
 });
